+++ CastlevaniaUit/InvisibleObject.h
@@ -33,7 +33,7 @@ public:
 	virtual void GetBoundingBox(float &x, float &y, float &framew, float &frameh){
 		x = this->x;
 		y = this->y;
-		framew = this->Width;
-		frameh = this->Height;
+		framew = (float)this->Width;
+		frameh = (float)this->Height;
 	}
 };
